// src/auth/dto/web3-login.dto.ts
// src/auth/dto/web3-login.dto.ts
import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
import {
  IsString,
  IsNotEmpty,
  IsEthereumAddress,
  IsOptional,
} from 'class-validator';

export class Web3LoginDto {
  @ApiProperty({
    description: 'Ethereum wallet address',
    example: '0x1234567890abcdef1234567890abcdef12345678',
  })
  @IsString()
  @IsNotEmpty()
  @IsEthereumAddress()
  walletAddress: string;

  @ApiProperty({ description: 'Message that was signed by the wallet' })
  @IsString()
  @IsNotEmpty()
  message: string;

  @ApiProperty({ description: 'Signature generated by signing the message' })
  @IsString()
  @IsNotEmpty()
  signature: string;

  @ApiPropertyOptional({
    description: 'The referrer id for the user',
  })
  @IsOptional()
  referrer?: string;
}